% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{myGibbsC}
\alias{myGibbsC}
\title{A Gibbs sampler using Rcpp}
\usage{
myGibbsC(N, b)
}
\arguments{
\item{N}{the number of samples}

\item{b}{the number of burn-in length}
}
\value{
a random sample of size \code{n}
}
\description{
A Gibbs sampler using Rcpp
}
\examples{
\dontrun{
XC <- myGibbsC(5000,1000)
}
}
