% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simudata.R
\name{simudata}
\alias{simudata}
\title{generate the desired simulation data}
\usage{
simudata(beta0, Beta, gamma0, Time, sig = 1, N, alpha)
}
\arguments{
\item{beta0}{Initial value of intercept item}

\item{Beta}{Initial value of network effect}

\item{gamma0}{Initial value of nodal effect}

\item{Time}{The observation time}

\item{sig}{Variance of the error term}

\item{N}{the number of nodes}

\item{alpha}{the parameter of power-law distribution}
}
\value{
a simulated dataset of N*(p+1)
}
\description{
Use network matrix, node-specific random vector, mean and covariance of the stationary state to generate the simulation data
}
\examples{
\dontrun{
N=50  #numbers of nodes
alpha=1.2  #parameter of power-law
Time=30   #The length of time
gamma0=c(-0.5,0.3,0.8,0,0) #fixed parameter of Z
beta0=rep(0.3,N) #Intercept term coefficient
Beta=c(0.1,0.5)  #Regression coefficient
Ysim=simudata(beta0, Beta, gamma0, Time, sig = 1, N, alpha)
est(Ysim,N,alpha,p)
}
}
