\name{StatComp21074-package}
\alias{StatComp21074-package}
\docType{package}
\title{
  Some simple functions to reproduce network vector autoregressions
}
\description{
  Simudata() is used to generate the data used in the simulation, and est() is used to calculate the estimator.
}
\author{
Yuxin Zhang

Maintainer: Yuxin Zhang <sa21204185@mail.ustc.edu.cn>
}
\keyword{ package }

